+++ const.py
@@ -1,3 +1,6 @@
 """Constants for the Flukso integration."""
 
 DOMAIN = "flukso"
+
+CONF_IGNORE_SENSORS = "ignore_sensors"
+DEFAULT_TIMEOUT = 10
